Move all action buttons to component
