+++ packages/app-js/package.json
@@ -13,5 +13,6 @@
     "@babel/runtime": "^7.2.0",
     "@polkadot/ui-app": "^0.23.34",
     "codeflask": "1.4.0"
-  }
+  },
+  "devDependencies": {}
 }
